requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-apscheduler==0.6.0
django-bootstrap-modal-forms==2.2.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-cuser==2017.3.16
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-localflavor==3.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.5
python-stdnum==1.17
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
whitenoise==5.3.0
